requirements.txt
mock
nose
pyyaml
